+++ include/Utils/StringOps.hpp
@@ -1,0 +1,11 @@
+#include <string>
+
+/**
+ * @brief Replaces all the ocurrences of a character in a string.
+ * 
+ * @param str The string.
+ * @param to_replace The character to replace.
+ * @param replacement The replacement of the character.
+ * @return std::string A copy of the string of text with the cars replaced.
+ */
+std::string replaceAllOcurrencesOfCharacter(std::string str, char to_replace, char replacement);
