+++ src/Utils/WindowOpsWindows.cpp
@@ -1,5 +1,5 @@
 #ifdef _WIN32
-#include "../include/OSUtils.hpp"
+#include "../../include/Utils/WindowOps.hpp"
 #include <Windows.h>
 
 void OS::maximizeWindow(sf::RenderWindow& window){
